move grinding and register test to MarketScenarioTest
